+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-
+## 0.2.0
+- additional output file: summary_report.txt, which contains ... summary report ;-)
+- simple test_app example
+- memory KB and MB output
 
 ## 0.1.0
  - initial version